Adsk contrib - Fix VC++ compilation warnings & some SonarCloud bugs (#996)
* Adsk contrib - Fix VC++ compilation warnings & some SonarCloud reported bugs


* Fix some doc warnings


* Fix macos build breaks


* Fix macos build breaks, part II


* Fix Windows build break: C++17 warning related to nodiscard


* Fix Windows build break: C++17 warning related to nodiscard, part II


* Fix Windows build break: C++17 warning related to nodiscard, part III


* Fix Windows build break: C++17 warning related to nodiscard, part IV


* C++ version detection improvment


* C++ version detection improvment, part II


* Fix Windows build break: C++17 warning related to nodiscard, part V


* Fix Windows build break: C++17 warning related to nodiscard, part VI


* Fix Windows build break: C++17 warning related to nodiscard, part VI

+++ src/bindings/python/PyLut1DTransform.cpp
@@ -77,7 +77,7 @@ void bindPyLut1DTransform(py::module & m)
 
                 self->setLength(size / 3);
 
-                for (size_t i = 0; i < size; i += 3)
+                for (unsigned long i = 0; i < size; i += 3)
                 {
                     self->setValue(i / 3, values[i], values[i+1], values[i+2]);
                 }
